Respond to PR review
